requirements.txt
discord.py
python-dotenv
google-generativeai
Pillow
aiohttp
better-profanity
fpdf2
